+++ tests/test_gamma_enhanced.cpp
@@ -611,7 +611,7 @@ TEST_F(GammaEnhancedTest, ParallelBatchPerformanceBenchmark) {
             gamma_dist.getCumulativeProbabilityWithStrategy(std::span<const double>(test_values), std::span<double>(cdf_results), libstats::performance::Strategy::SCALAR);
         }
         auto end = std::chrono::high_resolution_clock::now();
-        result.simd_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.simd_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 2. Standard Parallel Operations
         std::span<const double> input_span(test_values);
@@ -632,7 +632,7 @@ TEST_F(GammaEnhancedTest, ParallelBatchPerformanceBenchmark) {
             gamma_dist.getCumulativeProbabilityWithStrategy(input_span, cdf_output_span, libstats::performance::Strategy::PARALLEL_SIMD);
             end = std::chrono::high_resolution_clock::now();
         }
-        result.parallel_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.parallel_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 3. Work-Stealing Operations (use shared pool to avoid resource issues)
         if constexpr (requires {
@@ -677,7 +677,7 @@ TEST_F(GammaEnhancedTest, ParallelBatchPerformanceBenchmark) {
                 end = std::chrono::high_resolution_clock::now();
             }
         }
-        result.work_stealing_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.work_stealing_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 4. Cache-Aware Operations (use shared cache manager to avoid resource issues)
         if constexpr (requires {
@@ -716,7 +716,7 @@ TEST_F(GammaEnhancedTest, ParallelBatchPerformanceBenchmark) {
                 end = std::chrono::high_resolution_clock::now();
             }
         }
-        result.cache_aware_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.cache_aware_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // Calculate speedups
         result.parallel_speedup = result.simd_time_us > 0 ? static_cast<double>(result.simd_time_us) / static_cast<double>(result.parallel_time_us) : 0.0;
