+++ tests/test_gaussian_enhanced.cpp
@@ -533,7 +533,7 @@ TEST_F(GaussianEnhancedTest, ParallelBatchPerformanceBenchmark) {
             stdNormal.getCumulativeProbabilityWithStrategy(std::span<const double>(test_values), std::span<double>(cdf_results), libstats::performance::Strategy::SCALAR);
         }
         auto end = std::chrono::high_resolution_clock::now();
-        result.simd_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.simd_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 2. Standard Parallel Operations
         std::span<const double> input_span(test_values);
@@ -554,7 +554,7 @@ TEST_F(GaussianEnhancedTest, ParallelBatchPerformanceBenchmark) {
             stdNormal.getCumulativeProbabilityWithStrategy(input_span, cdf_output_span, libstats::performance::Strategy::PARALLEL_SIMD);
             end = std::chrono::high_resolution_clock::now();
         }
-        result.parallel_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.parallel_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 3. Work-Stealing Operations (use shared pool to avoid resource issues)
         if (op == "PDF") {
@@ -573,7 +573,7 @@ TEST_F(GaussianEnhancedTest, ParallelBatchPerformanceBenchmark) {
             stdNormal.getCumulativeProbabilityWithStrategy(input_span, cdf_output_span, libstats::performance::Strategy::WORK_STEALING);
             end = std::chrono::high_resolution_clock::now();
         }
-        result.work_stealing_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.work_stealing_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // 4. Cache-Aware Operations (use shared cache manager)
         if (op == "PDF") {
@@ -592,7 +592,7 @@ TEST_F(GaussianEnhancedTest, ParallelBatchPerformanceBenchmark) {
             stdNormal.getCumulativeProbabilityWithStrategy(input_span, cdf_output_span, libstats::performance::Strategy::CACHE_AWARE);
             end = std::chrono::high_resolution_clock::now();
         }
-        result.cache_aware_time_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+        result.cache_aware_time_us = static_cast<long>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
         
         // Calculate speedups
         result.parallel_speedup = result.simd_time_us > 0 ? static_cast<double>(result.simd_time_us) / static_cast<double>(result.parallel_time_us) : 0.0;
